requirements.txt
pdfplumber
pdf2image
pytesseract
pillow
